separate CATag.hpp from iGeom.h to enable linking to iGeom without having access to SDK headers
