+++ export_dagmc_cmd/igeom/iGeom.cpp
@@ -1,5 +1,6 @@
-#include "CubitEntity.hpp"
 #include "iGeom.h"
+#include "CATag.hpp"
+#include "CubitEntity.hpp"
 #include "CubitInterface.hpp"
 #include "GeometryQueryTool.hpp"
 #include "Body.hpp"
